Add read only cache interface
This adds a new method `RDo` on interface `Syncer` that is analog to
`Do` but instead of giving the function exclusive read write access it
allows concurrent read only access to the cache.

For this a new interface `ReadCacher` is passed to the function.
`ReadCacher` is implemented by every `Cacher`.